+++ test/test_mc.jl
@@ -116,9 +116,74 @@ end
 
 function Carlo.register_evaluables(
     ::Type{TestParallelRunMC},
-    eval::Evaluator,
+    eval::AbstractEvaluator,
     params::AbstractDict,
 )
     evaluate!((x, y) -> y - x^2, eval, :test4, (:test, :test2))
+    return nothing
+end
+
+mutable struct TestTemperedMC <: AbstractMC
+    μ::Float64
+    x::Float64
+end
+
+TestTemperedMC(params::AbstractDict) = TestTemperedMC(params[:μ], 0)
+
+Carlo.init!(mc::TestTemperedMC, ctx::MCContext, params::AbstractDict) = nothing
+function Carlo.sweep!(mc::TestTemperedMC, ctx::MCContext)
+    xnew = mc.x + 0.07 * randn(ctx.rng)
+    if rand(ctx.rng) < test_distribution(xnew, mc.μ) / test_distribution(mc.x, mc.μ)
+        mc.x = xnew
+    end
+end
+
+test_distribution(x, μ; f = 2) =
+    (exp(-(x - μ)^2 * f * (1 + μ^4)) + exp(-(x + μ)^2 * (1 + μ^4))) * 2 * (1 + μ^4) /
+    (sqrt(2π) * (1 + 1 / f))
+test_distribution_x(μ; f = 2) = (f^-0.5 - 1) / (f^-0.5 + 1) * μ
+test_distribution_x²(μ; f = 2) = (f^-1.5 + 1) / (f^-0.5 + 1) / (2 * (1 + μ^4)) + μ^2
+
+function Carlo.measure!(mc::TestTemperedMC, ctx::MCContext)
+    measure!(ctx, :X², mc.x^2)
+    measure!(ctx, :X, mc.x)
+    measure!(ctx, :XX², [mc.x, mc.x^2])
+
+    return nothing
+end
+
+function Carlo.write_checkpoint(mc::TestTemperedMC, out::HDF5.Group)
+    out["x"] = mc.x
+end
+
+function Carlo.read_checkpoint!(mc::TestTemperedMC, in::HDF5.Group)
+    mc.x = read(in, "x")
+end
+
+function Carlo.register_evaluables(
+    ::Type{TestTemperedMC},
+    eval::AbstractEvaluator,
+    params::AbstractDict,
+)
+    evaluate!((x, x², xx²) -> xx² .- [x, x²], eval, :Zero, (:X, :X², :XX²))
+    return nothing
+end
+
+function Carlo.parallel_tempering_log_weight_ratio(
+    mc::TestTemperedMC,
+    parameter_name::Symbol,
+    new_value,
+)
+    @assert parameter_name == :μ
+    return log(test_distribution(mc.x, new_value) / test_distribution(mc.x, mc.μ))
+end
+
+function Carlo.parallel_tempering_change_parameter!(
+    mc::TestTemperedMC,
+    parameter_name::Symbol,
+    new_value,
+)
+    @assert parameter_name == :μ
+    mc.μ = new_value
     return nothing
 end